refactor: magic strings and spellcheck in app
